Progress

+++ app/src/main/java/com/example/zufffinalyear/models/Incomeitem.kt
@@ -3,7 +3,7 @@ package com.example.zufffinalyear.models
 data class Incomeitem(
     val description: String = "",
     val tagNo: String = "", // Add tag number
-    val amount: Float = 0f,
+    val amount: Double = 0.0,
     val date: String = ""
 )
